@@ -1,9 +1,20 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Session\Fingerprint;
 
 use Psr\Http\Message\ServerRequestInterface;
-use Viserio\Component\Contract\Session\Fingerprint as FingerprintContract;
+use Viserio\Contract\Session\Fingerprint as FingerprintContract;
 
 class UserAgentGenerator implements FingerprintContract
 {
@@ -21,7 +32,7 @@ class UserAgentGenerator implements FingerprintContract
      */
     public function __construct(ServerRequestInterface $serverRequest)
     {
-        $userAgent    = '';
+        $userAgent = '';
         $serverParams = $serverRequest->getServerParams();
 
         if (isset($serverParams['REMOTE_ADDR'])) {
