@@ -1,18 +1,28 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Session\Handler;
 
 use Cake\Chronos\Chronos;
 use Psr\Http\Message\ServerRequestInterface;
-use Viserio\Component\Contract\Cookie\Cookie as CookieContract;
-use Viserio\Component\Contract\Cookie\QueueingFactory as JarContract;
+use Viserio\Contract\Cookie\QueueingFactory as JarContract;
 
 class CookieSessionHandler extends AbstractSessionHandler
 {
     /**
      * The cookie jar instance.
      *
-     * @var \Viserio\Component\Contract\Cookie\QueueingFactory
+     * @var \Viserio\Contract\Cookie\QueueingFactory
      */
     protected $cookie;
 
@@ -33,12 +43,12 @@ class CookieSessionHandler extends AbstractSessionHandler
     /**
      * Create a new cookie driven handler instance.
      *
-     * @param \Viserio\Component\Contract\Cookie\QueueingFactory $cookie
-     * @param int                                                $lifetime The session lifetime in seconds
+     * @param \Viserio\Contract\Cookie\QueueingFactory $cookie
+     * @param int                                      $lifetime The session lifetime in seconds
      */
     public function __construct(JarContract $cookie, int $lifetime)
     {
-        $this->cookie   = $cookie;
+        $this->cookie = $cookie;
         $this->lifetime = $lifetime;
     }
 
@@ -61,27 +71,37 @@ class CookieSessionHandler extends AbstractSessionHandler
     }
 
     /**
-     * {@inheritdoc}
+     * Cleanup old sessions.
+     *
+     * @see https://php.net/manual/en/sessionhandlerinterface.gc.php
+     *
+     * @param int $maxlifetime
+     *
+     * @return bool
      */
-    public function gc($lifetime): bool
+    public function gc($maxlifetime): bool
     {
         return true;
     }
 
     /**
-     * {@inheritdoc}
+     * Update timestamp of a session.
+     *
+     * @param string $sessionId   The session id
+     * @param string $sessionData
+     *
+     * @return bool
      */
-    public function updateTimestamp($sessionId, $data): bool
+    public function updateTimestamp($sessionId, $sessionData): bool
     {
         $cookies = $this->cookie->getQueuedCookies();
-        $cookie  = $cookies[$sessionId] ?? null;
+        $cookie = $cookies[$sessionId] ?? null;
 
         if ($cookie === null) {
             return false;
         }
 
         $this->cookie->queue($this->cookie->delete($sessionId));
-        // @var CookieContract $cookie
         $this->cookie->queue(
             $cookie->withExpires(
                 Chronos::now()->addSeconds($this->lifetime)->getTimestamp()
@@ -104,8 +124,8 @@ class CookieSessionHandler extends AbstractSessionHandler
 
         $decoded = \json_decode(\base64_decode($cookies[$sessionId], true), true);
 
-        if (\is_array($decoded) &&
-            (isset($decoded['expires']) && Chronos::now()->getTimestamp() <= $decoded['expires'])
+        if (\is_array($decoded)
+            && (isset($decoded['expires']) && Chronos::now()->getTimestamp() <= $decoded['expires'])
         ) {
             return $decoded['data'];
         }
@@ -122,7 +142,7 @@ class CookieSessionHandler extends AbstractSessionHandler
             $sessionId,
             \base64_encode(\json_encode(
                 [
-                    'data'    => $data,
+                    'data' => $data,
                     'expires' => Chronos::now()->addSeconds($this->lifetime)->getTimestamp(),
                 ],
                 \JSON_PRESERVE_ZERO_FRACTION
