@@ -1,5 +1,16 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Session\Handler;
 
 use SessionHandlerInterface;
@@ -46,7 +57,7 @@ class MigratingSessionHandler implements SessionHandlerInterface, SessionUpdateT
             $writeOnlyHandler = new StrictSessionHandler($writeOnlyHandler);
         }
 
-        $this->currentHandler   = $currentHandler;
+        $this->currentHandler = $currentHandler;
         $this->writeOnlyHandler = $writeOnlyHandler;
     }
 
@@ -75,7 +86,13 @@ class MigratingSessionHandler implements SessionHandlerInterface, SessionUpdateT
     }
 
     /**
-     * {@inheritdoc}
+     * Cleanup old sessions.
+     *
+     * @see https://php.net/manual/en/sessionhandlerinterface.gc.php
+     *
+     * @param int $maxlifetime
+     *
+     * @return bool
      */
     public function gc($maxlifetime): bool
     {
@@ -129,7 +146,12 @@ class MigratingSessionHandler implements SessionHandlerInterface, SessionUpdateT
     }
 
     /**
-     * {@inheritdoc}
+     * Update timestamp of a session.
+     *
+     * @param string $sessionId   The session id
+     * @param string $sessionData
+     *
+     * @return bool
      */
     public function updateTimestamp($sessionId, $sessionData): bool
     {
