@@ -1,5 +1,16 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Session\Tests\Handler;
 
 use PHPUnit\Framework\TestCase;
@@ -7,12 +18,12 @@ use Viserio\Component\Session\Handler\NullSessionHandler;
 
 /**
  * @internal
+ *
+ * @small
  */
 final class NullSessionHandlerTest extends TestCase
 {
-    /**
-     * @var \Viserio\Component\Session\Handler\NullSessionHandler
-     */
+    /** @var \Viserio\Component\Session\Handler\NullSessionHandler */
     private $handler;
 
     /**
@@ -27,41 +38,41 @@ final class NullSessionHandlerTest extends TestCase
 
     public function testOpen(): void
     {
-        $this->assertTrue($this->handler->open('/', 'test'));
+        self::assertTrue($this->handler->open('/', 'test'));
     }
 
     public function testClose(): void
     {
-        $this->assertTrue($this->handler->close());
+        self::assertTrue($this->handler->close());
     }
 
     public function testValidateId(): void
     {
-        $this->assertTrue($this->handler->validateId('test'));
+        self::assertTrue($this->handler->validateId('test'));
     }
 
     public function testUpdateTimestamp(): void
     {
-        $this->assertTrue($this->handler->updateTimestamp('test', ''));
+        self::assertTrue($this->handler->updateTimestamp('test', ''));
     }
 
     public function testGc(): void
     {
-        $this->assertTrue($this->handler->gc(100));
+        self::assertTrue($this->handler->gc(100));
     }
 
     public function testRead(): void
     {
-        $this->assertSame('', $this->handler->read('test'));
+        self::assertSame('', $this->handler->read('test'));
     }
 
     public function testWrite(): void
     {
-        $this->assertTrue($this->handler->write('test', ''));
+        self::assertTrue($this->handler->write('test', ''));
     }
 
     public function testDestroy(): void
     {
-        $this->assertTrue($this->handler->destroy('test'));
+        self::assertTrue($this->handler->destroy('test'));
     }
 }
